requirements.txt
dnspython==1.13.0